Refactor: Agent graph
Use the Agent graph to run the agent through different stages (nodes)

This makes it so that the context can be managed more easily and more complex flows can be designed
