+++ crates/but-bot/src/lib.rs
@@ -3,9 +3,11 @@ use but_tools::emit::Emitter;
 use gitbutler_command_context::CommandContext;
 use gitbutler_project::ProjectId;
 
+mod butbot;
 mod state;
 
-use crate::agent::{Agent, ButBot};
+use crate::agent::AgentGraph;
+use crate::butbot::ButBot;
 
 pub mod agent;
 
@@ -17,6 +19,7 @@ pub fn bot(
     openai: &OpenAiProvider,
     chat_messages: Vec<but_action::ChatMessage>,
 ) -> anyhow::Result<String> {
-    let mut but_bot = ButBot::new(ctx, emitter, message_id, project_id, openai);
-    but_bot.evaluate(chat_messages)
+    let mut but_bot = ButBot::new(ctx, emitter, message_id, project_id, openai, chat_messages);
+    let mut graph = AgentGraph::default();
+    graph.start(&mut but_bot)
 }
